+++ switch_model/balancing/unserved_load.py
@@ -6,6 +6,15 @@ Defines components to allow leaving some load unserved. This module is
 specially useful when running production costing simulations, though not
 strictly required in all cases.
 
+INPUT FILE FORMAT
+    The cost penalty of unserved load in units of $/MWh is the only parameter
+    that can be inputted. The following file is not mandatory, because the
+    parameter defaults to a value of 500 $/MWh. This file contains one header
+    row and one data row.
+
+    optional input files:
+        lost_load_cost.csv
+            unserved_load_penalty
 """
 
 import os
@@ -33,6 +42,7 @@ def define_components(mod):
 
     mod.unserved_load_penalty = Param(
         within=NonNegativeReals,
+        input_file="lost_load_cost.csv",
         default=500)
     mod.UnservedLoad = Var(
         mod.LOAD_ZONES, mod.TIMEPOINTS,
@@ -44,21 +54,3 @@ def define_components(mod):
         rule=lambda m, tp: sum(m.UnservedLoad[z, tp] *
             m.unserved_load_penalty for z in m.LOAD_ZONES))
     mod.Cost_Components_Per_TP.append('UnservedLoadPenalty')
-
-def load_inputs(mod, switch_data, inputs_dir):
-    """
-    The cost penalty of unserved load in units of $/MWh is the only parameter
-    that can be inputted. The following file is not mandatory, because the
-    parameter defaults to a value of 500 $/MWh. This file contains one header
-    row and one data row.
-
-    optional input files:
-        lost_load_cost.csv
-            unserved_load_penalty
-
-    """
-    switch_data.load_aug(
-        filename=os.path.join(inputs_dir, 'lost_load_cost.csv'),
-        optional=True, auto_select=True,
-        param=(mod.unserved_load_penalty,)
-    )
