+++ switch_model/solve.py
@@ -22,7 +22,7 @@ from switch_model.utilities import (
 )
 from switch_model.upgrade import do_inputs_need_upgrade, upgrade_inputs
 from switch_model.tools.graph.cli_graph import main as graph_main
-from switch_model.utilities.custom_loading import patch_to_allow_loading
+from switch_model.utilities.load_data import patch_to_allow_loading
 from switch_model.utilities.results_info import save_info, add_info, ResultsInfoSection
